requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
apturl==0.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.8.3
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
click-plugins==1.1.1
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
elasticsearch==7.9.1
exceptiongroup==1.1.2
fasteners==0.14.1
feedparser==6.0.10
frozenlist==1.4.0
future==0.18.2
git-filter-repo==2.34.0
gyp==0.1
h11==0.14.0
htmlwebshot==0.1.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
powerline-status==2.8.2
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
selenium==4.10.0
sgmllib3k==1.0.0
shodan==1.29.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
ssh-import-id==5.11
systemd-python==234
trio==0.22.1
trio-websocket==0.10.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wsproto==1.2.0
xdg==5
xkit==0.0.0
XlsxWriter==3.1.2
yara-python==4.3.1
yarl==1.9.2
zipp==1.0.0
